requirements.txt
# requirements.txt
# 专业交易系统依赖 - Python 3.12 兼容版本

# ============ 核心依赖 ============
ccxt>=4.0.0                 # 交易所连接
pandas>=2.0.0               # 数据处理
numpy>=1.24.0               # 数值计算
python-dotenv>=1.0.0        # 环境变量管理

# ============ 技术指标 ============
pandas-ta==0.4.71b0         # Pandas技术指标 (beta版本)

# ============ 数据存储 ============
sqlalchemy>=2.0.0           # 数据库ORM

# ============ 可视化 ============
matplotlib>=3.7.0
plotly>=5.15.0
mplfinance>=0.12.10b0

# ============ 机器学习 ============
scikit-learn>=1.3.0

# ============ 工具 ============
pyyaml>=6.0
loguru>=0.7.0               # 日志
requests>=2.31.0            # HTTP请求
websocket-client>=1.6.0     # WebSocket
schedule>=1.2.0             # 任务调度
aiohttp>=3.9.0              # 异步HTTP

# ============ 测试 ============
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# ============ 代码质量 ============
black>=23.0.0               # 代码格式化
flake8>=7.0.0               # 代码检查
